+++ CHANGELOG.md
@@ -11,7 +11,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ### Added
 - resource `pullzone_shield`: support Bot Categorization;
-- resource `pullzone_ratelimit_rule`: support response actions, counter keys and negated conditions;
+- resource `pullzone_ratelimit_rule`: support `condition.negated`, `limit.counter_key` and `response.action` ([#97](https://github.com/BunnyWay/terraform-provider-bunnynet/pull/97));
 
 ### Fixed
 - resource `pullzone`: unknown `permacache_storagezone` causes invalid plan ([#96](https://github.com/BunnyWay/terraform-provider-bunnynet/issues/96));